+++ lib/output/cache.go
@@ -29,8 +29,7 @@ func init() {
 		Summary: `
 Stores each message in a [cache](/docs/components/caches/about).`,
 		Description: `
-Caches are configured within the [resources section](/docs/components/caches/about)
-and can target any of the following types:
+Caches are configured as [resources](/docs/components/caches/about) and can target any of the following types:
 
 ` + cachesList + `
 The ` + "`target`" + ` field must point to a configured cache like follows:
@@ -41,13 +40,12 @@ output:
     target: foo
     key: ${!json("document.id")}
 
-resources:
-  caches:
-    foo:
-      memcached:
-        addresses:
-          - localhost:11211
-        ttl: 60
+resource_caches:
+  - label: foo
+    memcached:
+      addresses:
+        - localhost:11211
+      ttl: 60
 ` + "```" + `
 
 In order to create a unique ` + "`key`" + ` value per item you should use