+++ lib/processor/cache.go
@@ -22,13 +22,9 @@ func init() {
 			CategoryIntegration,
 		},
 		Summary: `
-Performs operations against a [cache resource](/docs/components/caches/about)
-for each message, allowing you to store or retrieve data within message payloads.`,
+Performs operations against a [cache resource](/docs/components/caches/about) for each message, allowing you to store or retrieve data within message payloads.`,
 		Description: `
-This processor will interpolate functions within the ` + "`key` and `value`" + `
-fields individually for each message. This allows you to specify dynamic keys
-and values based on the contents of the message payloads and metadata. You can
-find a list of functions [here](/docs/configuration/interpolation#bloblang-queries).`,
+This processor will interpolate functions within the ` + "`key` and `value`" + ` fields individually for each message. This allows you to specify dynamic keys and values based on the contents of the message payloads and metadata. You can find a list of functions [here](/docs/configuration/interpolation#bloblang-queries).`,
 		FieldSpecs: docs.FieldSpecs{
 			docs.FieldCommon("resource", "The [`cache` resource](/docs/components/caches/about) to target with this processor."),
 			docs.FieldDeprecated("cache"),
@@ -59,11 +55,10 @@ pipeline:
         value: "storeme"
     - bloblang: root = if errored() { deleted() }
 
-resources:
-  caches:
-    foocache:
-      redis:
-        url: tcp://TODO:6379
+resource_caches:
+  - label: foocache
+    redis:
+      url: tcp://TODO:6379
 `,
 			},
 			{
@@ -82,11 +77,10 @@ pipeline:
               key: '${! json("message.document_id") }'
         result_map: 'root.message.document = this'
 
-resources:
-  caches:
-    foocache:
-      memcached:
-        addresses: [ "TODO:11211" ]
+resource_caches:
+  - label: foocache
+    memcached:
+      addresses: [ "TODO:11211" ]
 `,
 			},
 		},